devide models
