+++ ebook_homebrew/models/models.py
@@ -1,0 +1,28 @@
+"""Provides Rest API models
+"""
+
+import datetime
+
+
+class UploadModel:
+    """Upload_id Model"""
+
+    def __init__(self, upload_id):
+        self.upload_id = str(upload_id)
+        self.release_date = datetime.datetime.now()
+
+
+class ErrorModel:
+    """Error Model"""
+
+    def __init__(self, error):
+        self.error = str(error)
+        self.errorDate = datetime.datetime.now()
+
+
+class FileNotFoundModel:
+    """FileNotFound Model"""
+
+    def __init__(self, reason):
+        self.reason = reason
+        self.errorDate = datetime.datetime.now()
