+++ setup.cfg
@@ -80,7 +80,7 @@ log_cli = false
 log_level = INFO
 log_file_level = INFO
 timeout = 5
-python_paths = ./ebook_homebrew ./ebook_homebrew/utils
+python_paths = ./ebook_homebrew ./ebook_homebrew/utils ./ebook_homebrew/models
 pep8ignore =
     E501
     doc_src/*.py ALL
